Support Custom Rendering For Pie Charts
I'd like to be able to animate a pie chart but for most animation frameworks, we need access to the children groups in order to modify the path parameters. By adding an optional render function, we can add animation to the pie chart while allowing users to choose their own animation framework.